chore(kumactl): split out kumactl test client so it can be reused in other projects
